refactor: movendo os filtros da pasta business para a de Services
